CON-1185: PoC Python enclave API using Jep
Using Jep to create a wrapper around the existing Java enclave API in Python. This has been achieved by implementing a normal Kotlin Enclave class, `PythonEnclaveAdapter`, which delegates the various calls into to Python code which is provided by the user as a script file. This “adapter” enclave is implemented in a new python-enclave-adapter module. Currently only mock mode works and it must be created manually via the `createMockHost` internal method.

The user will consume the Python API via duck typing. `PythonEnclaveAdapter` will look to see if the matching global function is defined and if so execute it:

* `on_enclave_startup()` - executed by `onStartup`
* `on_enclave_shutdown()` - executed by `onShutdown`
* `receive_from_untrusted_host(bytes)` - executed by `receiveFromUntrustedHost`. The Java byte array is converted to Python `bytes`. If there’s no return value then it is treated as null, otherwise the return value is expected to be `bytes`.
* `receive_enclave_mail(mail)` - executed by `receiveMail`. The Java `EnclaveMail` object is converted to a simpler Python version which is just a class holding the body, envelope and authenticated sender. The topic and sequence number are ignored for now. The authenticated sender is represented by its encoded binary form in `bytes`. This function is also much simpler than `receiveMail` in that the response can simply be returned from the function and it will encrypted as a response mail back to the sender key and posted.

There’s also a global `enclave_sign(data)` function available which delegates to the Conclave enclave’s signing ability via `signer()`.