CON-1192: Python API integrated with simulation mode
The user can now create a Python enclave by creating an enclave Gradle module with a single Python script in `src/main/python` (everything else in the Gradle project stays the same). The plugin will detect this and will take a fat jar of the `python-enclave-adapter` module, the Python script, the generated Gramine manifest (which is now created dynamically based on the local environment), and will bundle them into a zip file. This zip file is what the host will scan for when it loads an enclave. Any future data that needs to be part of a Gramine enclave can be put into this zip bundle.

The enclave scanning process was been rewritten to support this zip Gramine “bundle”, alongwith the existing .so GraalVM “bundle”. The Gramine enclave process will pick up this optional Python file and manually set it on the `PythonEnclaveAdapter` object.

Note, the integration test currently hangs on TC though it passes if run separately. It is only testing a very basic Python enclave, but instead should be updated to use something like PyTorch.