CON-1192: Python API integrated with simulation mode
The user can now create a Python enclave by creating an enclave Gradle module with a single Python script in `src/main/python` (everything else in the Gradle project stays the same). The plugin will detect this and will take a fat jar of the `python-enclave-adapter` module, the Python script, the generated Gramine manifest (which is now created dynamically based on the local environment), and will bundle them into a zip file. This zip file is what the host will scan for when it loads an enclave. Any future data that needs to be part of a Gramine enclave can be put into this zip bundle as well.

The enclave scanning process was been rewritten to support this zip Gramine “bundle”, alongwith the existing .so GraalVM “bundle”. The Gramine enclave process will pick up this optional Python file and manually set it on the `PythonEnclaveAdapter` object.

The `runtimeType` enclave config has been made less strict and will now ignore case. For Python enclaves it will default to Gramine, rejecting any GraalVM config.

A new python enclave has been added to the integration tests and which does a very basic Python test. This should be updated to something more substantial such as PyTorch. The integration tests had to be cleaned up to ensure it was running, including fixing the docker container so that Gramine can run inside it. This now means the Gramine integration tests no longer hang, but for some reason the Python test does not execute, even though it passes when run directly.