CON-1187: Fix QA tests by changing the name of "enclave-bundle" jars
QA tests are currently failing because this is where we use a different technique to deploy our host applications.
In QA we generate call `distZip`, `distTar` tasks in order to bundle the host, upload it to the CRAFT VMs and unzip it.
We have a bug when the host had multiple enclaves: we use the same name for different "enclave bundle" jars. Even though these jars are actually present in differently named directories, when we call `distZip` we actually take them out of the "directory structure" context and therefore they are conflicting.
This causes issues in QA: Gradle complains about a missing strategy to handle this. 
Changing the name as we do in this PR should solve this problem