CON-1178: Integrate Gramine SGX attestation
Here we introduce attestation for the Gramine flow.
An effort has been done in order to keep the hierarchy of classes as clean as possible with the one we currently use for Native Image/Graal.
In Gramine attestation works differently from Native Image/Graal flow.
In Native Image:

1. the host requests the "target info" (through SGX calls) and pass it to the enclave;
2. the enclave uses the "target info" to create a report, this is done with an in-enclave SGX call;
3. the enclave triggers another call to the host, asking it to generate a signed quote (here we have a recursive call);
4. the host returns the signed quote to the enclave, concluding the recursive call here above;
5. the enclave returns the signed quote back to the host
6. the host attests the signed quote calling the provisioning enclave (SGX calls)

In Gramine the flow is a bit simpler

1. the host asks for a signed quote to the enclave
2. the enclave create a report. Internally the "target info" is automatically retrieved by Gramine (by passing an empty array to the pseudo filesystem `/dev/attestation/target_info`, similarly to what we do in Native with the call `sgx_get_target_info`, where an empty array is passed in. The only difference is that in Gramine this is happening in the enclave)
3. the enclave prepares for us, automatically, a signed quote, by doing the quoting in background. There is a call to the AESM service for this.
4. the enclave returns the signed quote to the host (same as in Native Image)
5. the host attests the signed quote (same as in Native Image)

Note that the `initializeQuote` function is returning an empty `SgxTargetInfo`, which is the one passed to the enclave to create the report later on. We could also skip this step, and even remove that class and move the `Native.initQuoteDCAP` in the Attestation Service; in such a case we would need to directly pass the empty array in the `GramineEnclaveEnvironment->getSignedQuote->createReport`. I kept everything as it is just for consistency with the current code, but this can change in the futur5e.