CON-1223 - Updated the default values of maxThreads and the enclave_worker_threads.
The default value of `enclave_worker_threads` is set to 10.

Gramine's maximum threads will no longer be twice the value set in the `maxThreads` but instead the exact value provided in the `maxThreads` configuration field.

`maxThreads` default value has been set to 100. This value should cover most of the use-cases and it does not degrade the enclave performance nor does increases the memory usage of an enclave.

Removed the `maxThreads` configuration field from `integration-tests/python-enclave/enclave/build.gradle`

The docs were updated to include the new default value set for the maxThreads.