CON-1240: Fix EnclaveInfo integration test under Gramine
This test has been updated to work under both GraalVM and Gramine. We used to have native code that read the sections of the .so file to extract out the enclave metadata. This is no longer needed and has been removed. Instead the sgx_sign tool is used to extract out the `SIGSTRUCT` data structure. Gramine writes this structure out in the .sig file. This isn't generated by gramine-direct in simulation mode, so the manifest file is checked instead.

`SgxMetadataEnclaveCss`, which represents the `SIGSTRUCT`, has been renamed to `SgxEnclaveCss` to better reflect the naming in the SGX SDK code.

The generateEnclaveMetadata task has also been improved to use the `SIGSTRUCT` dump from sgx_sign, rather than the fiddly text output.