CON-1235: Ensure plugin tasks are "incremental"
Ensure the Gradle plugin tasks are “incremental”, i.e. they run if any of their inputs change and do not run if not. This property is vital to ensure security sensitive information such as the `productID` and `revocationLevel` are correctly applied to the enclave if the user changes them in the enclave’s build.gradle file (and Gradle `clean` isn't run). It also gives better developer experience by ensuring tasks are not run needlessly.

Most of the tasks now have extensive integration tests to ensure that they respond correctly to their inputs. The two Gramine manifest generation tasks had to be merged into a single one to make it incremental, and also to make it easier to test.

Also, the default path of the external key’s `signingMaterial` file is now all lowercase.