CON-1235: Fix issue with Gradle checking for input files are may not exist
Certain input file parameters in the enclave's Gradle config may not exist when the build process begins. These include `mrsignerSignature` and `signingMaterial`. These particular files are only created in the first-pass of the `externalKey` signing process (when `generateEnclaveSigningMaterial` is invoked), and their existence is only required in the second-pass.

The previous changes to make sure all the plugin tasks are "incremental" incorrectly marked the `EnclaveExtension` config objects as inputs to the `GenerateEnclaveProperties` task. `GenerateEnclaveProperties` only needs the properties of the parent `ConclaveExtension` object as inputs. To fix this, the `release`, `debug`, and `simulation` properties are now marked as `@Internal` and not `@Nested`.