CON-1178: Fix certificate issues of AESM service
With the previous changes we have introduced the AESM service. This works but it is currently throwing an exception related to missing certificates. The root cause of the problem is that AESM service needs the Azure DCAP Client in order to retrieve such certificates and therefore we need to add it to the related AESM container.
This PR adds that.

In order to make an integration test pass (`MockEnclaveEnvironmentHardwareCompatibilityTest`), we also have to increase the `maxThread` value of the enclave used in that test