CON-1203: Gradle plugin using conclave-build container for GraalVM on Linux
Originally, enclaves were built in the docker container only on macOS and Windows. Now, all GraalVM enclaves are built in the `conclave-build` container regardless the OS. Gramine enclaves are still built outside this container, therefore python, Gramine, etc still do need to be installed on the machines.

Docker is now installed in the integration-tests docker container. Conclave-build container is now run with the same user as the integration-tests container - this is to prevent permission issues when deleting files.