CON-1069 - Update Gradle plugin to pull conclave-build image
By default, the end user builds its own `conclave-build`.

The configuration property `useInternalDockerRepo` was added to the Conclave's Gradle configuration block. This property allows the integration tests to pull the `conclave-build` docker image. By default, the value of the new property is false.

Updated the integration tests to fetch the docker container `conclave-build`.