CON-1285: Allow back the usage of signing key outside of enclaves
In previous PRs we have forced the signing keys to be defined inside the enclave directories. This was due to Gramine tests that needed such change.
In this PR we reverted back to the previous behaviour, so that we can now use signing key files which are outside the enclave paths.