CON-1261: Remove usage of Docker container when running Java enclaves
In this patch we remove the need of using Docker to run Java enclaves.
When using Java and Gramine only few libraries are required and therefore we can narrow down the scope of the system libraries that gets used. In order to have a minimal set of Java libraries we also use JLink, which allows the creation of a customized runtime image.